Re-work to use backtrace() system calls.
Use the backtrace() and backtrace_symbols() system calls rather
than manually specifying function names. Also drop support for
dynamic flub allocation. This dramatically simplifies the API
for the end-users and simplifies the backend implementation.
It is also hoped that the information supplied by official system
calls will be more useful than what the end-user is both willing
and able to provide at any given point.

The new API is not compatible with the old API, but I'm the only
person who uses this (I think) so far, so, who cares, right?