Remove 'inline' declarations from header file.
Apparently the 'inline' keyword is only local to a translation unit;
therefore, if one wants an inline function to by inline in an external
file that is including the '.h' file, one would have to include the
'inline' keyword -and- the code in the '.h' file. Despite making sense
from a compilation perspective, this makes me unhappy, since it blurs
the line regarding the purpose of separate '.h' and '.c' files.

This patch fixes an issue that prevents linking with the library.